fix: useFeature top level hooks, tile source provider is seperate file
