+++ src/components/AzureMapFeature/useFeature.ts
@@ -9,28 +9,29 @@ export const useFeature = (
   shapeRef: ShapeType | null,
   featureRef: FeatureType | null
 ) => {
-  if(dataSourceRef instanceof atlas.source.DataSource){
-    // Simple feature's usecases and methods
-    useCheckRef<atlas.source.DataSource, FeatureType>(dataSourceRef, featureRef, (dref, fref) => {
+  // Simple feature's usecases and methods
+  useCheckRef<DataSourceType, FeatureType>(dataSourceRef, featureRef, (dref, fref) => {
+    if(dref instanceof atlas.source.DataSource){
       dref.add(fref)
       return () => {
         dref.remove(fref)
       }
-    })
+    } else if (dataSourceRef instanceof atlas.source.VectorTileSource) {
+      console.error(`Unable to add Feature(${fref.id}) to VectorTileSource(${dataSourceRef.getId()}): AzureMapFeature has to be a child of AzureMapDataSourceProvider`)
+    }
+  })
 
-    // Shape's usecases and methods
-    useCheckRef<atlas.source.DataSource, ShapeType>(dataSourceRef, shapeRef, (dref, sref) => {
+  // Shape's usecases and methods
+  useCheckRef<DataSourceType, ShapeType>(dataSourceRef, shapeRef, (dref, sref) => {
+    if(dref instanceof atlas.source.DataSource){
       dref.add(sref)
       return () => {
         dref.remove(sref)
       }
-    })
-  } else if (dataSourceRef instanceof atlas.source.VectorTileSource) {
-    useCheckRef<FeatureType, FeatureType>(featureRef, featureRef, (fref) => 
-      //NOTE: IAzureVectorTileSourceChildren won't allow adding AzureMapFeature as a child, still notify in case type check was suppresed
-      console.error(`Unable to add Feature(${fref.id}) to VectorTileSource(${dataSourceRef.getId()}): AzureMapFeature has to be a child of AzureMapDataSourceProvider`)
-    )
-  }
+    } else if (dataSourceRef instanceof atlas.source.VectorTileSource) {
+      console.error(`Unable to add Shape(${sref.getId()}) to VectorTileSource(${dataSourceRef.getId()}): AzureMapFeature has to be a child of AzureMapDataSourceProvider`)
+    }
+  })
 
   useEffect(() => {
     if (shapeRef && setCoords) {
