+++ src/contexts/AzureMapDataSourceContext.test.tsx
@@ -5,10 +5,9 @@ import React from 'react'
 import { AzureMapsContext } from '../contexts/AzureMapContext'
 import {
   AzureMapDataSourceProvider,
-  AzureMapVectorTileSourceProvider,
   AzureMapDataSourceContext
 } from '../contexts/AzureMapDataSourceContext'
-import { IAzureDataSourceStatefulProviderProps, IAzureVectorTileSourceStatefulProviderProps } from '../types'
+import { IAzureDataSourceStatefulProviderProps } from '../types'
 
 const mapContextProps = {
   mapRef: null,
@@ -40,54 +39,6 @@ const wrapWithDataSourceContext = (props: IAzureDataSourceStatefulProviderProps)
     </AzureMapsContext.Provider>
   )
 }
-
-const wrapWithVectorTileSourceContext = (props: IAzureVectorTileSourceStatefulProviderProps) => ({
-  children
-}: {
-  children?: any
-}) => {
-  return (
-    <AzureMapsContext.Provider
-      value={{
-        ...mapContextProps,
-        mapRef
-      }}
-    >
-      <AzureMapVectorTileSourceProvider {...{ ...props }}>{children}</AzureMapVectorTileSourceProvider>
-    </AzureMapsContext.Provider>
-  )
-}
-
-describe('AzureMapVectorTileSourceProvider tests', () => {
-  it('should create data source with passed id and options', () => {
-    const { result } = renderHook(() => useContextConsumer(), {
-      wrapper: wrapWithVectorTileSourceContext({ id: 'id', options: { minZoom: 0, maxZoom: 12 } })
-    })
-
-    expect(result.current.dataSourceRef?.getId()).toEqual('id')
-    expect(result.current.dataSourceRef?.getOptions()).toEqual({ minZoom: 0, maxZoom: 12 })
-  })
-
-  it('should add data source to the map ref on mount', () => {
-    mapRef.sources.add = jest.fn()
-    const { result } = renderHook(() => useContextConsumer(), {
-      wrapper: wrapWithVectorTileSourceContext({ id: 'id' })
-    })
-    expect(mapRef.sources.add).toHaveBeenCalledWith(result.current.dataSourceRef)
-  })
-
-  it('should add event to data source', () => {
-    mapRef.events.add = jest.fn()
-    renderHook(() => useContextConsumer(), {
-      wrapper: wrapWithVectorTileSourceContext({ id: 'id', events: { sourceadded: (source) => {} } })
-    })
-    expect(mapRef.events.add).toHaveBeenCalledWith(
-      'sourceadded',
-      expect.any(Object),
-      expect.any(Function)
-    )
-  })
-})
 
 describe('AzureMapDataSourceProvider tests', () => {
   it('should add data source to the map ref on mount', () => {
